+++ test/SampleElement.m
@@ -1,5 +1,5 @@
 %% Element Design
-Element.type = 4;                   % Rectangular piston
+Element.type = int32(4);            % Rectangular piston
 Element.params_m = [lambda/2; ...   % Element face width, m
                     lambda/4; ...   % Element face height, m
                     0];
